:wrench: chore: fixed inconsistent naming
