--- app/src/main/java/shub39/momentum/data/createNotificationChannel.kt
@@ -1,32 +1,0 @@
-/*
- * Copyright (C) 2026  Shubham Gorai
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <https://www.gnu.org/licenses/>.
- */
-package shub39.momentum.data
-
-import android.app.NotificationChannel
-import android.app.NotificationManager
-import android.content.Context
-
-fun createNotificationChannel(context: Context) {
-    val name = "Momentum"
-    val descriptionText = "Reminders for montages"
-    val importance = NotificationManager.IMPORTANCE_DEFAULT
-    val channel = NotificationChannel("1", name, importance).apply { description = descriptionText }
-    val notificationManager: NotificationManager =
-        context.getSystemService(Context.NOTIFICATION_SERVICE) as NotificationManager
-
-    notificationManager.createNotificationChannel(channel)
-}
