--- app/src/main/java/shub39/momentum/data/getPlaceholder.kt
@@ -1,32 +1,0 @@
-/*
- * Copyright (C) 2026  Shubham Gorai
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <https://www.gnu.org/licenses/>.
- */
-package shub39.momentum.data
-
-import android.graphics.Bitmap
-import android.graphics.Bitmap.createBitmap
-import android.graphics.Color
-import androidx.compose.ui.graphics.asImageBitmap
-import androidx.compose.ui.graphics.painter.BitmapPainter
-import androidx.compose.ui.graphics.painter.Painter
-
-fun getPlaceholder(): Painter {
-    return BitmapPainter(
-        createBitmap(1, 1, Bitmap.Config.ARGB_8888)
-            .apply { eraseColor(Color.GREEN) }
-            .asImageBitmap()
-    )
-}
