+++ common/core/src/main/kotlin/shub39/momentum/core/EnumExt.kt
